almost done

+++ client/src/reducers/control.js
@@ -14,7 +14,11 @@ const initialState = {
 export default (state = initialState, { type, payload }) => {
   switch (type) {
   case LOGIN_SUCCESS:
-    localStorage.setItem('token' , payload.token);
+
+    if (payload.keepUserLoggedIn){
+      localStorage.setItem('token' , payload.token);
+    }
+
     return{
       ...state,
       authToken:payload.token,
